# swagger-spec.json
{"openapi":"3.0.0","paths":{"/user/{userId}":{"get":{"description":"특정 사용자의 정보를 조회합니다.","operationId":"UserController_getUser","parameters":[{"name":"userId","required":true,"in":"path","description":"사용자 ID","schema":{"type":"number"}}],"responses":{"200":{"description":"사용자 정보 조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserRespDto"}}}},"404":{"description":"사용자를 찾을 수 없음"}},"summary":"사용자 조회","tags":["User"]}},"/user":{"post":{"description":"새로운 사용자를 생성합니다.","operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"사용자 생성 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRespDto"}}}}},"summary":"사용자 생성","tags":["User"]}},"/point/charge":{"patch":{"operationId":"PointController_chargePoint","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChargePointReqDto"}}}},"responses":{"200":{"description":""}},"tags":["Point"]}},"/point/{userId}":{"get":{"operationId":"PointController_getPoint","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Point"]}},"/product":{"get":{"operationId":"ProductController_getProducts","parameters":[],"responses":{"200":{"description":""}},"tags":["Product"]}},"/coupon/user/{userId}":{"get":{"operationId":"CouponController_getCouponList","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Coupon"]}},"/coupon/issue/{userId}/{couponId}":{"post":{"operationId":"CouponController_issueCoupon","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"couponId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["Coupon"]}}},"info":{"title":"E-Commerce API","description":"E-Commerce API 명세서","version":"1.0","contact":{}},"tags":[{"name":"Point","description":""},{"name":"Coupon","description":""},{"name":"Product","description":""},{"name":"Payment","description":""},{"name":"Order","description":""}],"servers":[],"components":{"schemas":{"GetUserRespDto":{"type":"object","properties":{"userId":{"type":"number","example":1,"description":"사용자 ID"},"points":{"type":"number","example":10000,"description":"보유 포인트"},"createdAt":{"format":"date-time","type":"string","example":"2024-03-14T12:00:00Z","description":"생성 일시"},"updatedAt":{"format":"date-time","type":"string","example":"2024-03-14T12:00:00Z","description":"수정 일시"}},"required":["userId","points","createdAt","updatedAt"]},"CreateUserDto":{"type":"object","properties":{}},"CreateUserRespDto":{"type":"object","properties":{"userId":{"type":"number","example":1,"description":"사용자 ID"},"points":{"type":"number","example":10000,"description":"보유 포인트"},"createdAt":{"format":"date-time","type":"string","example":"2024-03-14T12:00:00Z","description":"생성 일시"}},"required":["userId","points","createdAt"]},"ChargePointReqDto":{"type":"object","properties":{}}}}}